Refactor UDP protocol layers
